requirements.txt
streamlit==1.18.0
torch==2.1.0
torchvision==0.15.0
Pillow==9.6.0
numpy==1.24.3
